% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvlrt-object-S3-methods.R
\name{all.equal.pvlrt}
\alias{all.equal.pvlrt}
\title{Test if two pvlrt objects are (Nearly) Equal}
\usage{
\method{all.equal}{pvlrt}(target, current, ...)
}
\arguments{
\item{target}{First pvlrt object (output of \link{pvlrt}).}

\item{current}{Second pvlrt object.}

\item{...}{Arguments passed to \link{all.equal.default}.}
}
\description{
Test if two pvlrt objects are (Nearly) Equal
}
\details{
Compares all values and attributes of target and current \code{pvlrt} objects except running times.
See \link{all.equal.default} for details on the generic function.
}
\seealso{
all.equal.default
}
