% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvlrt-object-S3-methods.R
\name{as.matrix.pvlrt}
\alias{as.matrix.pvlrt}
\title{Casting a \code{pvlrt} object as a matrix of log LR statistics}
\usage{
\method{as.matrix}{pvlrt}(x, ...)
}
\arguments{
\item{x}{a \code{pvlrt} object; an output of function \code{pvlrt}().}

\item{...}{other input parameters. Currently unused.}
}
\description{
Casting a \code{pvlrt} object as a matrix of log LR statistics
}
