% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt_extended_vanilla_poisson.R
\name{ext_lr_test_vanilla_poisson}
\alias{ext_lr_test_vanilla_poisson}
\title{Pseudo Likelihood Ratio Test for determining significant AE-Drug pairs under
zero-inflated Poisson model}
\usage{
ext_lr_test_vanilla_poisson(
  contin_table,
  drug_class_idx,
  nsim = 10000,
  fisher_p_level = 1e-04,
  return_p.value = FALSE,
  ...
)
}
\arguments{
\item{contin_table}{IxJ contingency table showing pairwise counts of adverse effects
for I AE and J Drugs}

\item{drug_class_idx}{a list, with the h-th entry providing the h-th group/class of drugs.
By default, each drug forms its own class. If more than one drug is present in a class, an
extended LRT is performed. See examples.}

\item{nsim}{Number of simulated null contingency table to use for computing the p-value of the test}

\item{...}{Other arguments. Currently unused.}

\item{omega_est_vec}{vector (for all drugs) of estimates of the zero-inflation.
If NULL, then estimated from the data under a Gamma process assumption.}
}
\description{
Pseudo Likelihood Ratio Test for determining significant AE-Drug pairs under
zero-inflated Poisson model
}
