% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvlrt-postprocessing-utils.R
\name{extract_AE_names}
\alias{extract_AE_names}
\alias{extract_Drug_names}
\title{Extracting AE and Drug names from a pvlrt object}
\usage{
extract_AE_names(object)

extract_Drug_names(object)
}
\description{
Extracting AE and Drug names from a pvlrt object
}
