% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{faers22q3raw}
\alias{faers22q3raw}
\title{FDA FAERS dataset for 2022 Q3}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 496312 rows and 3 columns.
}
\source{
\url{https://fis.fda.gov/extensions/FPD-QDE-FAERS/FPD-QDE-FAERS.html}
}
\usage{
faers22q3raw
}
\description{
The raw FDA FAERS dataset for 2022 Q3; downloaded from FDA's website
and then subsetted to inlcude incidences with \code{ROLE_COD == "PS"}.
}
\details{
obtained from the FDA FAERS database for the
quarters 2014Q3 - 2020Q4

This is a raw incidence data stored as a data.table with each row corresponding
to a specific incidence index by case ids. It contains the following
columns:
\itemize{
\item \code{CASEID} - case ids for each incidence
\item \code{DRUG} - names of the drugs
\item \code{AE} - names of the adverse events
}
}
\keyword{datasets}
