% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gbca}
\alias{gbca}
\title{FDA GBCA dataset with all observed 1707 adverse effects}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 1707 rows and 10 columns.
}
\source{
\url{https://fis.fda.gov/extensions/FPD-QDE-FAERS/FPD-QDE-FAERS.html}
}
\usage{
gbca
}
\description{
FAERS 2014Q3 - 2020Q4
}
\details{
Includes 6 GBCA's:

gadobenate, gadobutrol, gadodiamide, gadofosveset, gadopentetate, gadoterate,
gadoteridol, gadoversetamide, gadoxetate

Corresponding to all 1707 observed adverse effects (PTs) observed in statins
}
\keyword{datasets}
