% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvlrt-plots.R
\name{heatmap_pvlrt}
\alias{heatmap_pvlrt}
\title{Heatmap showing LR test results}
\usage{
heatmap_pvlrt(
  object,
  nrow = NULL,
  ncol = NULL,
  color_by = "p.value",
  show_n = TRUE,
  show_pvalue = FALSE,
  show_lrstat = FALSE,
  arrange_alphabatical = FALSE,
  digits = 2,
  ...
)
}
\arguments{
\item{object}{pvlrt object}

\item{...}{additional arguments passed to pheatmap::pheatmap()}
}
\description{
Heatmap showing LR test results
}
\examples{
test1 <- pvlrt(statin46)
heatmap_pvlrt(test1)
heatmap_pvlrt(test1, arrange_alphabetical = TRUE)

}
