% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvlrt-wrappers.R
\name{lrt_poisson}
\alias{lrt_poisson}
\alias{lrt_vanilla_poisson}
\title{Likelihood Ratio Test under the (vanilla, non-zero-inflated) Poisson model}
\usage{
lrt_poisson(contin_table, nsim = 10000, parametrization = "rrr", ...)

lrt_vanilla_poisson(contin_table, nsim = 10000, parametrization = "rrr", ...)
}
\arguments{
\item{contin_table}{IxJ contingency table showing pairwise counts of adverse
effects for I AE and J Drugs}

\item{nsim}{Number of simulated null contingency table to use for computing
the p-value of the test}

\item{...}{additional arguments passed to pvlrt}
}
\description{
Likelihood Ratio Test under the (vanilla, non-zero-inflated) Poisson model
}
\note{
\code{lrt_poisson()} and \code{lrt_vanilla_poisson()}
are both wrappers for \code{pvlrt()} with
\code{omega_vec = rep(0, ncol(contin_table))}
}
\examples{

data("lovastatin")

# no grouping -- each drug forms its own class
test1 <- lrt_poisson(lovastatin)

}
\seealso{
\link{pvlrt}
}
