% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt_vanilla_poisson.R
\name{lrt_vanilla_poisson}
\alias{lrt_vanilla_poisson}
\title{Likelihood Ratio Test for determining significant AE-Drug pairs}
\usage{
lrt_vanilla_poisson(
  contin_table,
  nsim = 10000,
  drug_class_idx = as.list(1:ncol(contin_table)),
  test_drug_idx = 1:ncol(contin_table),
  ...
)
}
\arguments{
\item{contin_table}{IxJ contingency table showing pairwise counts of adverse effects
for I AE and J Drugs}

\item{nsim}{Number of simulated null contingency table to use for computing the p-value of the test}

\item{drug_class_idx}{a list, with the h-th entry providing the h-th group/class of drugs.
By default, each drug forms its own class. If more than one drug is present in a class, an
extended LRT is performed. See examples.}

\item{...}{Other arguments. Currently unused.}
}
\description{
Likelihood Ratio Test for determining significant AE-Drug pairs
}
\examples{

data("lovastatin")

# no grouping -- each drug forms its own class
test1 <- lrt_vanilla_poisson(lovastatin)
## extract the observed LRT statistic
attr(test1, "lrstat")


# grouped drugs --
# group1 : drug 1, drug 2
# group 2: drug 3
drug_groups <- list(c(1, 2), 3)
test2 <- lrt_vanilla_poisson(lovastatin, drug_class_idx = drug_groups)

}
