% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvlrt-wrappers.R
\name{lrt_zi_poisson}
\alias{lrt_zi_poisson}
\title{Pseudo Likelihood Ratio Test under the
zero-inflated Poisson model with relative reporting rate parametrization}
\usage{
lrt_zi_poisson(contin_table, nsim = 10000, ...)
}
\arguments{
\item{contin_table}{IxJ contingency table showing pairwise counts of adverse
effects for I AE and J Drugs}

\item{nsim}{Number of simulated null contingency table to use for computing
the p-value of the test}

\item{...}{additional arguments passed to pvlrt}
}
\description{
Pseudo Likelihood Ratio Test under the
zero-inflated Poisson model with relative reporting rate parametrization
}
\note{
\code{lrt_zi_poisson()} is a wrapper for \code{pvlrt()} with
\code{parametrization = "rrr"}.
}
\examples{

data("lovastatin")
# no grouping -- each drug forms its own class
test1 <- lrt_zi_poisson(lovastatin)
## extract the observed LRT statistic
attr(test1, "lrstat")
## extract the estimated omegas
attr(test1, "omega")

# grouped drugs --
# group1 : drug 1, drug 2
# group 2: drug 3
drug_groups <- list(c(1, 2), 3)
test2 <- lrt_zi_poisson(lovastatin, drug_class_idx = drug_groups)

\dontrun{
# test omegas
test3 <- lrt_zi_poisson(lovastatin, test_omega = TRUE)
## extract the estimate, statistic, and p-values for omegas
attr(test3, "omega")
attr(test3, "omega_lrstat")
attr(test3, "omega_pvalue")


# use gamma prior assumption on signals while
# estimating omegas
test3 <- lrt_zi_poisson(
  lovastatin,
  use_gamma_smooth_omega = TRUE
)
attr(test3, "omega")
}

}
