% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvlrt-object-S3-methods.R
\name{plot.pvlrt}
\alias{plot.pvlrt}
\title{Plotting method for a pvlrt object}
\usage{
\method{plot}{pvlrt}(object, type = "heatmap", ...)
}
\arguments{
\item{object}{a \code{pvlrt} object, which is the output of the function
\link{pvlrt} or one of its wrappers such as \link{lrt_zi_poisson},
\link{lrt_poisson} and \code{lrt_vanilla_poisson}.}

\item{type}{character string determining the type of plot to show.
Available choices are "heatmap" (default) which calls \link{heatmap_pvlrt},
and "barplot" which calls \link{barplot.pvlrt} with the additional arguments
supplied in ...}

\item{...}{additional arguments passed to heatmap_pvlrt or barplot.pvlrt
depending on \code{type}.}
}
\description{
Plotting method for a pvlrt object
}
