% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvlrt-object-S3-methods.R
\name{print.pvlrt}
\alias{print.pvlrt}
\title{Print method for pvlrt objects}
\usage{
\method{print}{pvlrt}(object, significance_level = 0.05, topn = 12, digits = 2, ...)
}
\arguments{
\item{object}{a \code{pvlrt} object, which is the output of the function
\link{pvlrt} or one of its wrappers such as \link{lrt_zi_poisson},
\link{lrt_poisson} and \code{lrt_vanilla_poisson}.}

\item{significance_level}{numeric. Level of significance.}

\item{topn}{number of top (with respect to likelihood ratio statistic value)
pairs to show at the given significance level.}

\item{digits}{number of digits to show after the decimal place.}

\item{...}{other input parameters. Currently unused.}
}
\description{
Print method for pvlrt objects
}
