% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvLRT-package.R
\name{pvLRT-package}
\alias{pvLRT-package}
\title{\code{pvLRT}: An R package implementing various Likelihood Ratio Test-based
approaches to pharmacovigilance}
\description{
\code{pvLRT} is an R package that implements a suite of
likelihood ratio test based methods to use in pharmacovigilance.
It can handle adverse effects data on several simultaneous drugs,
with possibly zero inflated report counts. Several testing and post-processing
functions are implemented.
}
