% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvlrt-function.R
\name{pvlrt}
\alias{pvlrt}
\title{Pseudo Likelihood Ratio Test for determining significant AE-Drug pairs under
Poisson and zero-inflated Poisson models}
\usage{
pvlrt(
  contin_table,
  nsim = 10000,
  omega_vec = NULL,
  zi_prob = NULL,
  no_zi_idx = NULL,
  drug_class_idx = as.list(1:ncol(contin_table)),
  test_drug_idx = 1:ncol(contin_table),
  grouped_omega_est = FALSE,
  test_zi = NULL,
  test_omega = NULL,
  pval_ineq_strict = FALSE,
  parametrization = "rrr",
  null_boot_type = "parametric",
  is_zi_structural = TRUE,
  ...
)
}
\arguments{
\item{contin_table}{IxJ contingency table showing pairwise counts of adverse
effects for I AE and J Drugs}

\item{nsim}{Number of simulated null contingency table to use for computing
the p-value of the test}

\item{zi_prob, omega_vec}{vector (for all drugs) of estimates of the zero-inflation.
If NULL, then is estimated from the data. See also the
description of the argument \code{grouped_omega_est}. If \code{omega_vec = rep(0, ncol(contin_table))},
then test reduces to an ordinary (non-zero inflated) Poisson test. NOTE: \code{zi_prob} and \code{omega_vec}
are alias.}

\item{no_zi_idx}{List of pairs {(i, j)} where zero inflation is not allowed. To
specify the entirety i-th row (or j-th column) use c(i, 0) (or c(0, j)). See examples.}

\item{drug_class_idx}{a list, with the h-th entry providing the h-th group/class
of drugs. By default, each drug forms its own class. If more than one drug is
present in a class, an extended LRT is performed. See examples.}

\item{test_drug_idx}{integer vector representing the columns (drug indices) of contin_table to be tested for signal.
Defaults to all columns.}

\item{grouped_omega_est}{Logical. When performing a test with grouped drug classes (extended LRT),
should the estimated zero-inflation parameter "omega" reflect
the corresponding grouping? If TRUE, then the estimated omegas are obtained by combining
columns from the same group, and if FALSE (default), then omegas are estimated separately for each drug (column)
irrespective of the groups specified through  \code{drug_class_idx}. Ignored if \code{omega_vec} is
supplied/non-\code{NULL} (i.e., not estimated).}

\item{test_zi, test_omega}{logical indicators specifying whether to perform a
pseudo likelihood ratio test for zero inflation. Defaults to FALSE. Ignored
if \code{omega_vec} is supplied (is non-NULL). Defaults to FALSE.
NOTE: \code{test_omega} and \code{test_zi} are aliases.}

\item{pval_ineq_strict}{logical. Use a strict inequality in the definition of the p-values?  Defaults to FALSE.}

\item{is_zi_structural}{logical. Do the inflated zeros correspond to structural
zeros (indicating impossible AE-Drug combination)? This determines how the
bootstrap null zero-inflation indicators are generated. If TRUE (default),
then then the null zero-inflation random indicators are randomly generated using the
(estimated) \emph{conditional} probabilities of zero inflation given observed
counts. If FALSE, then they are generated using the \emph{marginal} (drug-specific)
estimated probabilities of zero-inflation.}

\item{use_gamma_smooth_omega}{logical. Use a gamma prior (smoothing) on the signals (lambdas) while estimating
omega from data. Defaults to FALSE. NOTE: if TRUE, then the gamma prior produces a marginal negative binomial
distribution for the counts, which is then optimized for estimating omega.}
}
\description{
Pseudo Likelihood Ratio Test for determining significant AE-Drug pairs under
Poisson and zero-inflated Poisson models
}
\note{
\code{lrt_poisson} is a wrapper function on \code{lrt_zi_poisson} with \code{omega_vec = rep(0, ncol(contin_table))}
}
\examples{

data("statin46")
# no grouping -- each drug forms its own class,
test1 <- pvlrt(statin46)
test1
## extract the observed LRT statistic
extract_lrstat_matrix(test1)
## extract the estimated omegas
extract_zi_probability(test1)

# grouped drugs --
# group 1: drug 1, drug 2
# group 2: drug 3, drug 4
# drug 5, 6, 7 in their own groups
drug_groups <- list(c(1, 2), c(3, 4), 5, 6, 7)
test2 <- pvlrt(statin46, drug_class_idx = drug_groups)
test2

# specify no zero inflation at the entirety of the last row and the last column
no_zi_idx <- list(c(nrow(statin46), 0), c(0, ncol(statin46)))
test3 <- pvlrt(statin46, no_zi_idx = no_zi_idx)
test3

# use non-parametric bootstrap to evaluate the null distribution
test4 <- pvlrt(statin46, null_boot_type = "non-parametric")
test4

\dontrun{
# test zi probabilities (omegas)
test5 <- pvlrt(statin46, test_omega = TRUE)
test5
}

}
