% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvLRTextra-package.R
\name{pvLRTextra-package}
\alias{pvLRTextra-package}
\title{\code{pvLRTextra}: An R package implementing various Likelihood Ratio Test-based
approaches to pharmacovigilance and more}
\description{
\code{pvLRTextra} is an R package that extends the functionality of pvLRT in various ways.
This is a developmental package, and stable functionalities will be ported to pvLRT.
}
