% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rv}
\alias{rv}
\title{FDA rotavirus vaccine dataset with 794 adverse events observed among
combined old (age >= 1 year) and  young (age < 1 year) individuals}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 794 rows and 2 columns.
}
\source{
\url{https://vaers.hhs.gov/data/datasets.html}
}
\usage{
rv
}
\description{
A vaccine-Adverse event (AE) count dataset (contingency table)
obtained from the FDA VAERS database for the year 1999
}
\details{
Data are stored in the forms of contingency table, with the
vaccines listed across the columns and the 794 AEs presented across
the rows. Each cell in the contingency table represents the total
report counts associated with that vaccine/AE pair and detected
in the FDA VAERS database for the year 1999.

The dataset contains two columns -- one for the rotavirus vaccine, and another
for other (37 vaccines combined).
}
\keyword{datasets}
