% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rvold}
\alias{rvold}
\title{FDA rotavirus vaccine dataset with 727 adverse events observed among
"old" (non-infant; age >= 1 year) individuals}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 727 rows and 2 columns.

An object of class \code{matrix} (inherits from \code{array}) with 727 rows and 2 columns.
}
\source{
\url{https://vaers.hhs.gov/data/datasets.html}

\url{https://vaers.hhs.gov/data/datasets.html}
}
\usage{
rvold

rvold
}
\description{
A vaccine-Adverse event (AE) count dataset (contingency table)
obtained from the FDA VAERS database for the year 1999

A vaccine-Adverse event (AE) count dataset (contingency table)
obtained from the FDA VAERS database for the year 1999
}
\details{
Data are stored in the forms of contingency table, with the
vaccines listed across the columns and the 727 AEs presented across
the rows. Each cell in the contingency table represents the total
report counts (from "old" individuals with age >= 1 year) associated
with that vaccine/AE pair and detected
in the FDA VAERS database for the year 1999.

The dataset contains two columns -- one for the rotavirus vaccine, and another
for other (37 vaccines combined).

Data are stored in the forms of contingency table, with the
vaccines listed across the columns and the 346 AEs presented across
the rows. Each cell in the contingency table represents the total
report counts from young individuals with age < 1 year associated
with that vaccine/AE pair and detected
in the FDA VAERS database for the year 1999.

The dataset contains two columns -- one for the rotavirus vaccine, and another
for other (37 vaccines combined).
}
\keyword{datasets}
