% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{statin}
\alias{statin}
\title{FDA Statin dataset with 6039 adverse events}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 6039 rows and 7 columns.
}
\source{
\url{https://fis.fda.gov/extensions/FPD-QDE-FAERS/FPD-QDE-FAERS.html}
}
\usage{
statin
}
\description{
A Drug-Adverse event (AE) count dataset (contingency table)
obtained from the FDA FAERS database for the
quarters 2014Q3 - 2020Q4
}
\details{
Data are stored in the form of a contingency table, with
drugs listed across the columns and the 6039 AEs presented across
the rows. Each cell in the contingency table represents the total
report counts associated with that drug/AE pair and detected
in the FDA FAERS database during 2014Q3 - 2020Q4.

The dataset catalogs 6 statin drugs (across columns):

Atorvastatin, Fluvastatin, Lovastatin, Pravastatin, Rosuvastatin, Simvastatin

Corresponding to all 6039 observed adverse events (AEs) observed in statins
}
\keyword{datasets}
