% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvlrt-postprocessing-utils.R
\name{extract_AE_names}
\alias{extract_AE_names}
\alias{extract_Drug_names}
\alias{set_AE_names}
\alias{set_Drug_names}
\title{Extracting and setting AE and Drug names from a pvlrt object}
\usage{
extract_AE_names(object)

extract_Drug_names(object)

set_AE_names(object, old, new)

set_Drug_names(object, old, new)
}
\arguments{
\item{object}{a \code{pvlrt} object, which is the output of the function
\link{pvlrt} or one of its wrappers such as \link{lrt_zi_poisson},
\link{lrt_poisson} and \code{lrt_vanilla_poisson}.}

\item{old}{character vector containing the old names}

\item{new}{character vector containing the new names}
}
\description{
Extracting and setting AE and Drug names from a pvlrt object
}
