% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvlrt-postprocessing-utils.R
\name{extract_lrstat_matrix}
\alias{extract_lrstat_matrix}
\alias{extract_p_value_matrix}
\alias{extract_zi_probability}
\alias{extract_significant_pairs}
\title{Extract various summary measures from a pvlrt object}
\usage{
extract_lrstat_matrix(object, ...)

extract_p_value_matrix(object, ...)

extract_zi_probability(object, ...)

extract_significant_pairs(object, significance_level = 0.05, ...)
}
\arguments{
\item{object}{a \code{pvlrt} object, which is the output of the function
\link{pvlrt} or one of its wrappers such as \link{lrt_zi_poisson},
\link{lrt_poisson} and \code{lrt_vanilla_poisson}.}

\item{...}{other input parameters. Currently unused.}

\item{significance_level}{numeric. Level of significance.}
}
\description{
Extract various summary measures from a pvlrt object
}
