% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gbca}
\alias{gbca}
\title{FDA GBCA dataset with all observed 1707 adverse events}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 1707 rows and 10 columns.
}
\source{
\url{https://fis.fda.gov/extensions/FPD-QDE-FAERS/FPD-QDE-FAERS.html}
}
\usage{
gbca
}
\description{
A Drug-Adverse event (AE) count dataset (contingency table)
obtained from the FDA FAERS database for the
quarters 2014Q3 - 2020Q4
}
\details{
Data are stored in the form of a contingency table, with
drugs listed across the columns and the 1707 AEs presented across
the rows. Each cell in the contingency table represents the total
report counts associated with that drug/AE pair and detected
in the FDA FAERS database during 2014Q3 - 2020Q4.

The dataset contains 6 Gadolinium-Based Contrast Agents (GBCAs) as drugs:

gadobenate, gadobutrol, gadodiamide, gadofosveset, gadopentetate, gadoterate,
gadoteridol, gadoversetamide, gadoxetate

Corresponding to all 1707 observed adverse events (AEs) as curated in FAERS database.
}
\keyword{datasets}
