% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvlrt-plots.R
\name{heatmap_pvlrt}
\alias{heatmap_pvlrt}
\alias{barplot.pvlrt}
\title{Heatmap showing LR test results}
\usage{
heatmap_pvlrt(
  object,
  AE = NULL,
  Drug = NULL,
  grep = FALSE,
  fill_measure = "p_value",
  show_n = TRUE,
  arrange_alphabetical = FALSE,
  show_p_value = FALSE,
  show_lrstat = FALSE,
  p_value_lower = 0,
  p_value_upper = 1,
  lrstat_lower = 0,
  lrstat_upper = Inf,
  n_lower = 0,
  n_upper = Inf,
  remove_outside = FALSE,
  digits = 2,
  border_color = "white",
  ...
)

\method{barplot}{pvlrt}(
  height,
  AE = NULL,
  Drug = NULL,
  grep = FALSE,
  x_axis_measure = "lrstat",
  fill_measure = "p_value",
  show_n = FALSE,
  arrange_alphabetical = FALSE,
  show_p_value = FALSE,
  show_lrstat = FALSE,
  p_value_lower = 0,
  p_value_upper = 1,
  lrstat_lower = 0,
  lrstat_upper = Inf,
  n_lower = 0,
  n_upper = Inf,
  remove_outside = FALSE,
  digits = 2,
  Drug_nrow = 1,
  x_axis_trans = FALSE,
  x_axis_trans_fn = "log1p",
  border_color = "white",
  ...
)
}
\arguments{
\item{object}{pvlrt object}

\item{fill_measure}{Measure to govern the filling}

\item{...}{additional arguments passed to pheatmap::pheatmap()}

\item{height}{a \code{pvlrt} object; output of \code{pvlrt}().}
}
\description{
Heatmap showing LR test results
}
\examples{
test1 <- pvlrt(statin46)
heatmap_pvlrt(test1)
heatmap_pvlrt(test1, arrange_alphabetical = TRUE)

}
