% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lovastatin}
\alias{lovastatin}
\title{FDA lovastatin dataset}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 47 rows and 3 columns.
}
\source{
\url{https://fis.fda.gov/extensions/FPD-QDE-FAERS/FPD-QDE-FAERS.html}
}
\usage{
lovastatin
}
\description{
Lovastatin Drug-AE dataset (contingency table) obtained from the FAERS
databases for the quarters 2014Q3 - 2020Q4
}
\details{
Drugs are stored across columns, and the 46 AEs are stored across the rows

Includes 1 column for lovastatin, and one combined column for all other drugs.
}
\keyword{datasets}
