% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lovastatin}
\alias{lovastatin}
\title{FDA lovastatin dataset}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 47 rows and 3 columns.
}
\source{
\url{https://fis.fda.gov/extensions/FPD-QDE-FAERS/FPD-QDE-FAERS.html}
}
\usage{
lovastatin
}
\description{
A Drug-Adverse event (AE) count dataset (contingency table)
obtained from the FDA FAERS database for the
quarters 2014Q3 - 2020Q4
}
\details{
Data are stored in the form of a contingency table, with
drugs listed across the columns and the 46 AEs presented across
the rows. Each cell in the contingency table represents the total
report counts associated with that drug/AE pair and detected
in the FDA FAERS database during 2014Q3 - 2020Q4.

The dataset contains 1 column for the lovastatin drug,
and one column for all other drugs combined.
}
\keyword{datasets}
