% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt_vanilla_poisson.R
\name{lrt_vanilla_poisson}
\alias{lrt_vanilla_poisson}
\title{Likelihood Ratio Test for determining significant AE-Drug pairs}
\usage{
lrt_vanilla_poisson(contin_table, nsim = 10000, ...)
}
\arguments{
\item{contin_table}{IxJ contingency table showing pairwise counts of adverse effects
for I AE and J Drugs}

\item{nsim}{Number of simulated contin_table to use for computing the p-value of the test}
}
\description{
Likelihood Ratio Test for determining significant AE-Drug pairs
}
