% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt_zi_poisson.R
\name{lrt_zi_poisson}
\alias{lrt_zi_poisson}
\title{Pseudo Likelihood Ratio Test for determining significant AE-Drug pairs under
zero-inflated Poisson model}
\usage{
lrt_zi_poisson(contin_table, nsim = 10000, omega_est_vec = NULL, ...)
}
\arguments{
\item{contin_table}{IxJ contingency table showing pairwise counts of adverse effects
for I AE and J Drugs}

\item{nsim}{Number of simulated contin_table to use for computing the p-value of the test}

\item{omega_est_vec}{vector (for all drugs) of estimates of the zero-inflation.
If NULL, then estimated from the data under a Gamma process assumption.}
}
\description{
Pseudo Likelihood Ratio Test for determining significant AE-Drug pairs under
zero-inflated Poisson model
}
