% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvlrt-object-S3-methods.R
\name{plot.pvlrt}
\alias{plot.pvlrt}
\title{Plotting method for a pvlrt object}
\usage{
\method{plot}{pvlrt}(x, type = "heatmap", ...)
}
\arguments{
\item{x}{a \code{pvlrt} object; an output of function \code{pvlrt}().}

\item{type}{character string determining the type of plot to show.
Available choices are "heatmap" (default) which calls \link{heatmap_pvlrt},
and "barplot" which calls \link{barplot.pvlrt} with the additional arguments
supplied in ...}

\item{...}{additional arguments passed to heatmap_pvlrt or barplot.pvlrt
depending on \code{type}.}
}
\description{
Plotting method for a pvlrt object
}
