% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvlrt-function.R
\name{pvlrt}
\alias{pvlrt}
\title{Pseudo Likelihood Ratio Test for determining significant AE-Drug pairs under
zero-inflated Poisson model}
\usage{
pvlrt(
  contin_table,
  nsim = 10000,
  omega_vec = NULL,
  zi_prob = NULL,
  drug_class_idx = as.list(1:ncol(contin_table)),
  test_drug_idx = 1:ncol(contin_table),
  grouped_omega_est = FALSE,
  test_zi = FALSE,
  test_omega = test_zi,
  pval_ineq_strict = FALSE,
  parametrization = "rrr",
  ...
)
}
\arguments{
\item{contin_table}{IxJ contingency table showing pairwise counts of adverse
effects for I AE and J Drugs}

\item{nsim}{Number of simulated null contingency table to use for computing
the p-value of the test}

\item{zi_prob, omega_vec}{vector (for all drugs) of estimates of the zero-inflation.
If NULL, then is estimated from the data. See also the
description of the argument \code{grouped_omega_est}. If \code{omega_vec = rep(0, ncol(contin_table))},
then test reduces to an ordinary (non-zero inflated) Poisson test. NOTE: \code{zi_prob} and \code{omega_vec}
are alias.}

\item{drug_class_idx}{a list, with the h-th entry providing the h-th group/class
of drugs. By default, each drug forms its own class. If more than one drug is
present in a class, an extended LRT is performed. See examples.}

\item{test_drug_idx}{integer vector representing the columns (drug indices) of contin_table to be tested for signal.
Defaults to all columns.}

\item{grouped_omega_est}{Logical. When performing a test with grouped drug classes (extended LRT),
should the estimated zero-inflation parameter "omega" reflect
the corresponding grouping? If TRUE, then the estimated omegas are obtained by combining
columns from the same group, and if FALSE (default), then omegas are estimated separately for each drug (column)
irrespective of the groups specified through  \code{drug_class_idx}. Ignored if \code{omega_vec} is
supplied/non-\code{NULL} (i.e., not estimated).}

\item{test_zi, test_omega}{logical indicators to specifiy whether to perform a bootstrap pseudo likelihood ratio test for omega? Ignored if
\code{omega_vec} is supplied (is non-NULL). Defaults to FALSE. NOTE: \code{test_omega} and
\code{test_zi} are aliases.}

\item{pval_ineq_strict}{logical. Use a strict inequality in the definition of the p-values?  Defaults to FALSE.}

\item{use_gamma_smooth_omega}{logical. Use a gamma prior (smoothing) on the signals (lambdas) while estimating
omega from data. Defaults to FALSE. NOTE: if TRUE, then the gamma prior produces a marginal negative binomial
distribution for the counts, which is then optimized for estimating omega.}
}
\description{
Pseudo Likelihood Ratio Test for determining significant AE-Drug pairs under
zero-inflated Poisson model
}
\note{
\code{lrt_poisson} is a wrapper function on \code{lrt_zi_poisson} with \code{omega_vec = rep(0, ncol(contin_table))}
}
\examples{



data("lovastatin")
# no grouping -- each drug forms its own class
test1 <- pvlrt(lovastatin)
## extract the observed LRT statistic
attr(test1, "lrstat")
## extract the estimated omegas
attr(test1, "omega")

# grouped drugs --
# group1 : drug 1, drug 2
# group 2: drug 3
drug_groups <- list(c(1, 2), 3)
test2 <- lrt_zi_poisson(lovastatin, drug_class_idx = drug_groups)

\dontrun{
# test omegas
test3 <- lrt_zi_poisson(lovastatin, test_omega = TRUE)
## extract the estimate, statistic, and p-values for omegas
attr(test3, "omega")
attr(test3, "omega_lrstat")
attr(test3, "omega_pvalue")


# use gamma prior assumption on signals while
# estimating omegas
test3 <- lrt_zi_poisson(
  lovastatin,
  use_gamma_smooth_omega = TRUE
)
attr(test3, "omega")
}

}
