% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{statin}
\alias{statin}
\title{FDA Statin dataset with all observed 6039 adverse effects}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 6039 rows and 7 columns.
}
\source{
\url{https://fis.fda.gov/extensions/FPD-QDE-FAERS/FPD-QDE-FAERS.html}
}
\usage{
statin
}
\description{
FAERS 2014Q3 - 2020Q4
}
\details{
Includes 6 Statins:

Atorvastatin, Fluvastatin, Lovastatin, Pravastatin, Rosuvastatin, Simvastatin

Corresponding to all 6039 observed adverse effects (PTs) observed in statins
}
\keyword{datasets}
