% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{statin1491}
\alias{statin1491}
\title{FDA Statin dataset with 1491 adverse effects}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 1491 rows and 7 columns.
}
\source{
\url{https://fis.fda.gov/extensions/FPD-QDE-FAERS/FPD-QDE-FAERS.html}
}
\usage{
statin1491
}
\description{
Drug-AE dataset (contingency table) obtained from the FAERS
databases for the quarters 2014Q3 - 2020Q4
}
\details{
Drugs are stored across columns, and the 1491 AEs are stored across the rows

the dataset includes 6 statin drugs:

Atorvastatin, Fluvastatin, Lovastatin, Pravastatin, Rosuvastatin, Simvastatin

Corresponding to 1491 adverse effects (PTs) common between statins and GBCA
}
\keyword{datasets}
