% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{statin46}
\alias{statin46}
\title{FDA Statin dataset with 46 adverse events}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 47 rows and 7 columns.
}
\source{
\url{https://fis.fda.gov/extensions/FPD-QDE-FAERS/FPD-QDE-FAERS.html}
}
\usage{
statin46
}
\description{
A Drug-Adverse event (AE) count dataset (contingency table)
obtained from the FAERS databases for the
quarters 2014Q3 - 2020Q4
}
\details{
Data are stored in the form of a contingency table, with
drugs listed across the columns and the 46 AEs presented across
the rows. Each cell in the contingency table represents the total
report counts associated with that drug/AE pair and detected
in the FDA FAERS database during 2014Q3 - 2020Q4.

The dataset catalogs 6 statin drugs (across columns):

Atorvastatin, Fluvastatin, Lovastatin, Pravastatin, Rosuvastatin, Simvastatin

The 46 adverse events presented across the rows are considered
significant by FDA.
}
\seealso{
\link{statin}, \link{statin1491}, \link{gbca}
}
\keyword{datasets}
