% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvlrt-object-S3-methods.R
\name{summary.pvlrt}
\alias{summary.pvlrt}
\title{Summary method for a pvlrt object}
\usage{
\method{summary}{pvlrt}(object, show_zi = FALSE, ...)
}
\arguments{
\item{object}{a \code{pvlrt} object, which is the output of the function
\link{pvlrt} or one of its wrappers such as \link{lrt_zi_poisson},
\link{lrt_poisson} and \code{lrt_vanilla_poisson}.}

\item{show_zi}{logical. Should summary of the estimates and tests
(if performed) of the zero inflation parameters be returned?
Defaults to FALSE. If TRUE, then the zero inflation summary is included
as an attribute with name "zi". See examples.}

\item{...}{other input parameters. Currently unused.}
}
\description{
Summary method for a pvlrt object
}
\examples{
\dontrun{
test1 <- pvlrt(statin46, test_zi = TRUE)
summary(test1)
tmp <- summary(test1, show_zi = TRUE)
attr(tmp, "zi")
}
}
